backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
arxiv==2.3.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
daphne==4.2.1
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
feedparser==6.0.12
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jiter==0.12.0
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.7
numpy==2.2.6
openai==2.8.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
RapidFuzz==3.14.3
redis==7.1.0
requests==2.32.5
seaborn==0.13.2
service-identity==24.2.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
django-environ==0.12.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-extensions==4.1
django-redis==6.0.0
celery==5.5.3
pandas==2.3.3
black==25.11.0
isort==7.0.0
flake8==7.3.0
pylint==4.0.2
coverage==7.12.0
psycopg2-binary==2.9.10
django-filter==25.2.0